review changes. introduced ContentStoreIntegrationTest.
